New folder - Copy/requirements.txt
# Full-stack Python web app (Streamlit UI + analysis services)
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
folium>=0.15.0
streamlit-folium>=0.18.0
plotly>=5.18.0
python-whois>=0.8.0
dnspython>=2.4.0
cryptography>=41.0.0
pyopenssl>=23.0.0
fpdf2>=2.7.0
python-docx>=1.1.0
